#72 BillingAddressTrait and ShippingAddressTrait
